shared template updates
