+++ tests/Amusoft.Templates.Tests/Amusoft.Templates.Tests.csproj
@@ -19,7 +19,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Amusoft.DotnetNew.Tests" Version="0.2.0-alpha.91" />
+		<PackageReference Include="Amusoft.DotnetNew.Tests" Version="0.2.0-alpha.93" />
 		<PackageReference Include="Amusoft.XUnit.NLog.Extensions" Version="2.4.0"/>
 		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.9.4"/>
 		<PackageReference Include="NLog" Version="4.7.11"/>
