+++ tests/Amusoft.Templates.Tests/Cases/DotNetLibraryRepoTests.cs
@@ -115,14 +115,14 @@ namespace Amusoft.Templates.Tests.Cases
 
 				await scaffold.RestoreAsync($"src/{sourceName}.sln", null, CancellationToken.None);
 				await scaffold.BuildAsync($"src/{sourceName}.sln", null, CancellationToken.None);
+				await scaffold.TestAsync($"src/{sourceName}.sln", null, CancellationToken.None);
 				await Verifier
 					.Verify(logging.ToFullString(PrintKind.All))
 					.UseParameters(sourceName);
 			}
-			catch (BuildFailedException e)
+			catch (CliException e)
 			{
-				OutputHelper.WriteLine(e.Output);
-				return;
+				Assert.Fail(e.Output);
 			}
 		}
 	}
