+++ tests/Amusoft.Templates.Tests/Snapshots/DotNetLibraryRepoTests.BuildCheck_sourceName=GeneratedProject.verified.txt
@@ -18,7 +18,18 @@ success: restore "{Scaffold}/src/GeneratedProject.sln" -v m --ignore-failed-sour
     "{Scaffold}\src\GeneratedProject\bin\Debug\net6.0\GeneratedProject.dll",
     "{Scaffold}\src\GeneratedProject\bin\Debug\netstandard2.0\GeneratedProject.dll",
     "{Scaffold}\tests\GeneratedProject.IntegrationTests\bin\Debug\net6.0\GeneratedProject.IntegrationTests.dll",
+    "{Scaffold}\tests\GeneratedProject.Tests.Shared\bin\Debug\net6.0\GeneratedProject.Tests.Shared.dll",
     "{Scaffold}\tests\GeneratedProject.UnitTests\bin\Debug\net6.0\GeneratedProject.UnitTests.dll"
+  ]
+}
+
+---TestResult---
+
+{
+  "Command": "test {Scaffold}/src/GeneratedProject.sln -v m --no-restore",
+  "Lines": [
+    "Passed!  - Failed:     0, Passed:     1, Skipped:     0, Total:     1, Duration: < SCRUBBED - GeneratedProject.IntegrationTests.dll (net6.0)",
+    "Passed!  - Failed:     0, Passed:     1, Skipped:     0, Total:     1, Duration: < SCRUBBED - GeneratedProject.UnitTests.dll (net6.0)"
   ]
 }
