+++ tests/Amusoft.Templates.Tests/Snapshots/DotNetLibraryRepoTests.BuildCheck_sourceName=GeneratedProject2.verified.txt
@@ -18,7 +18,18 @@ success: restore "{Scaffold}/src/GeneratedProject2.sln" -v m --ignore-failed-sou
     "{Scaffold}\src\GeneratedProject2\bin\Debug\net6.0\GeneratedProject2.dll",
     "{Scaffold}\src\GeneratedProject2\bin\Debug\netstandard2.0\GeneratedProject2.dll",
     "{Scaffold}\tests\GeneratedProject2.IntegrationTests\bin\Debug\net6.0\GeneratedProject2.IntegrationTests.dll",
+    "{Scaffold}\tests\GeneratedProject2.Tests.Shared\bin\Debug\net6.0\GeneratedProject2.Tests.Shared.dll",
     "{Scaffold}\tests\GeneratedProject2.UnitTests\bin\Debug\net6.0\GeneratedProject2.UnitTests.dll"
+  ]
+}
+
+---TestResult---
+
+{
+  "Command": "test {Scaffold}/src/GeneratedProject2.sln -v m --no-restore",
+  "Lines": [
+    "Passed!  - Failed:     0, Passed:     1, Skipped:     0, Total:     1, Duration: < SCRUBBED - GeneratedProject2.IntegrationTests.dll (net6.0)",
+    "Passed!  - Failed:     0, Passed:     1, Skipped:     0, Total:     1, Duration: < SCRUBBED - GeneratedProject2.UnitTests.dll (net6.0)"
   ]
 }
