+++ tests/Amusoft.Templates.Tests/Snapshots/DotNetLibraryRepoTests.FileCheck_sourceName=GeneratedProject.verified.txt
@@ -1,0 +1,1 @@
+﻿
