+++ tests/Amusoft.Templates.Tests/Snapshots/DotNetLibraryRepoTests.FileCheck_sourceName=GeneratedProject2.verified.txt
@@ -1,0 +1,1 @@
+﻿
